backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & AI
openai==1.10.0
langchain==0.1.5
langchain-openai==0.0.5
tiktoken==0.5.2

# Vector Database
pinecone-client==3.0.3
# Alternative: FAISS for local development
faiss-cpu==1.7.4

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1
beanie==1.24.0  # MongoDB ODM

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.1

# PDF Generation
reportlab==4.0.9
pypdf2==3.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Linting & Formatting
black==23.12.1
ruff==0.1.11
mypy==1.8.0
